+++ modules/pgsqlroute.cpp
@@ -5,7 +5,7 @@
  * Postgres SQL based routing
  *
  * Yet Another Telephony Engine - a fully featured software PBX and IVR
- * Copyright (C) 2004 Null Team
+ * Copyright (C) 2004, 2005 Null Team
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -48,7 +48,7 @@ public:
 bool RouteHandler::received(Message &msg)
 {
     char buffer[2048];
-    unsigned long long tmr = Time::now();
+    u_int64_t tmr = Time::now();
     String called(msg.getValue("called"));
     if (called.null())
 	return false;
@@ -93,7 +93,7 @@ bool PrerouteHandler::received(Message &msg)
     char buffer[2048];
 //    char select_called[200];
 //    char select_channel[200];
-    unsigned long long tmr = Time::now();
+    u_int64_t tmr = Time::now();
     // return immediately if there is already a context
     if (msg.getValue("context"))
 	return false;
